requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
flask-cors==4.0.1
PyMySQL==1.1.1
